requirements.txt
langchain
langchain-openai
langchain-community
langchain-pinecone
beautifulsoup4
python-dotenv
streamlit
pinecone-client
